# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: GmbH API
  version: 1.2.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/units:
    get:
      tags:
        - Unit
      summary: Request all units
      responses:
        '200':
          description: Units list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitListResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Unit
      summary: Create one unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnitCreateRequest'
      responses:
        '200':
          description: Unit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/units/{id}:
    get:
      tags:
        - Unit
      summary: Request unit by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Unit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Unit
      summary: Update unit
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnitUpdateRequest'
      responses:
        '200':
          description: Unit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Unit
      summary: Delete unit
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Empty response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tables:
    get:
      tags:
        - Table
      summary: Request all tables
      responses:
        '200':
          description: Tables list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableListResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Table
      summary: Create one table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableCreateRequest'
      responses:
        '200':
          description: Table
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tables/{id}:
    get:
      tags:
        - Table
      summary: Request table by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Table
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Table
      summary: Update table
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableUpdateRequest'
      responses:
        '200':
          description: Table
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Table
      summary: Delete table
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Empty response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users:
    get:
      tags:
        - User
      summary: Request all users
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - User
      summary: Create one user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/{id}:
    get:
      tags:
        - User
      summary: Request user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - User
      summary: Update user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - User
      summary: Delete user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Empty response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-access-token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        errors:
          type: object
          properties:
            msg:
              type: string
          required:
            - msg
      required:
        - errors
...
    Unit:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    UnitResponse:
      type: object
      properties:
        unit:
          $ref: '#/components/schemas/Unit'
      required:
        - unit
    UnitListResponse:
      type: object
      properties:
        units:
          type: array
          items:
            $ref: '#/components/schemas/Unit'
      required:
        - units
    UnitCreateRequest:
      type: object
      properties:
        unit:
          $ref: '#/components/schemas/Unit'
      required:
        - unit
    UnitUpdateRequest:
      type: object
      properties:
        unit:
          $ref: '#/components/schemas/Unit'
      required:
        - unit
    Table:
      type: object
      properties:
        id:
          type: string
        areaId:
          type: string
        name:
          type: string
        x:
          type: number
        y:
          type: number
        custom:
          type: boolean
        enabled:
          type: boolean
    TableResponse:
      type: object
      properties:
        table:
          $ref: '#/components/schemas/Table'
      required:
        - table
    TableListResponse:
      type: object
      properties:
        tables:
          type: array
          items:
            $ref: '#/components/schemas/Table'
      required:
        - tables
    TableCreateRequest:
      type: object
      properties:
        table:
          $ref: '#/components/schemas/Table'
      required:
        - table
    TableUpdateRequest:
      type: object
      properties:
        table:
          $ref: '#/components/schemas/Table'
      required:
        - table
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        role:
          type: string
          enum: ['admin', 'waiter']
        areas:
          type: array
          items:
            type: string
        password:
          type: string
    UserResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
      required:
        - user
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
      required:
        - users
    UserCreateRequest:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
      required:
        - user
    UserUpdateRequest:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
      required:
        - user
...